requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.40.61
botocore==1.40.61
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.2.1
comm==0.2.3
datasets==4.8.2
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
filelock==3.25.2
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
ijson==3.5.0
inspect_ai==0.3.195
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
jiter==0.13.0
jmespath==1.1.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
linkify-it-py==2.1.0
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
multiprocess==0.70.19
nest-asyncio==1.6.0
nest-asyncio2==1.7.2
numpy==2.4.3
openai==2.28.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pathlib_abc==0.5.2
pexpect==4.9.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
s3fs==2025.9.0
s3transfer==0.14.0
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
srsly==2.5.2
stack-data==0.6.3
tenacity==9.1.4
textual==8.1.1
tiktoken==0.12.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
universal_pathlib==0.3.10
urllib3==2.6.3
wcwidth==0.6.0
widgetsnbextension==4.0.15
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zipfile-zstd==0.0.4
zipp==3.23.0
zstandard==0.25.0
